+++ paymentsheet/src/main/java/com/stripe/android/checkout/CheckoutControllerState.kt
@@ -4,8 +4,8 @@ import android.graphics.Bitmap
 import android.os.Bundle
 import android.os.Parcelable
 import com.stripe.android.checkout.CheckoutController.Session
-import com.stripe.android.checkout.ece.AvailableExpressButtonTypesFactory
 import com.stripe.android.common.model.CommonConfiguration
+import com.stripe.android.elements.ece.AvailableExpressButtonTypesFactory
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
