+++ paymentsheet/src/main/java/com/stripe/android/checkout/CheckoutSessionMappers.kt
@@ -3,7 +3,7 @@ package com.stripe.android.checkout
 import android.graphics.Bitmap
 import com.stripe.android.checkout.CheckoutController.Session
 import com.stripe.android.checkout.CheckoutController.Session.PaymentOptionDisplayData
-import com.stripe.android.checkout.ece.ExpressButtonType
+import com.stripe.android.elements.ece.ExpressButtonType
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentsheet.repositories.CheckoutSessionResponse
 import com.stripe.android.paymentsheet.verticalmode.CurrencySelectorOptionsFactory
