+++ paymentsheet/src/test/java/com/stripe/android/elements/ece/ExpressCheckoutElementInteractorStateFactory.kt
@@ -1,6 +1,6 @@
 @file:OptIn(CheckoutSessionPreview::class)
 
-package com.stripe.android.checkout.ece
+package com.stripe.android.elements.ece
 
 import com.stripe.android.checkout.GooglePayConfiguration
 import com.stripe.android.link.LinkAccountUpdate
